requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework-simplejwt==5.2.2
djangorestframework==3.14.0
pip==23.1.2
psycopg2==2.9.6
psycopg==3.1.8
PyJWT==2.6.0
pytz==2023.3
setuptools==65.5.1
sqlparse==0.4.4
typing-extensions==4.5.0
wheel==0.38.4
